Automatically index and sign on build
Add new flag --generate-index (defaulting to true)
which will generate the index after build.

In addition, if --signing-key is provided, proceed
to sign the index after it is generated.

Also moved index and signing code into internal/sign/
in order to prevent circular dependency.

Lastly, added examples/minimal.yaml to have a
melange build to quickly test these things with.
