+++ .gitignore
@@ -16,3 +16,6 @@
 .vscode/*
 
 melange
+melange.rsa
+melange.rsa.pub
+packages/
