+++ examples/minimal.yaml
@@ -1,0 +1,15 @@
+package:
+  name: minimal
+  version: 0.0.1
+  description: a very basic melange example
+environment:
+  contents:
+    repositories:
+      - https://dl-cdn.alpinelinux.org/alpine/edge/main
+    packages:
+      - alpine-baselayout-data
+      - busybox
+pipeline:
+  - runs: |
+      mkdir -p "${{targets.destdir}}"
+      echo "hello" > "${{targets.destdir}}/hello"
