+++ pkg/cli/build.go
@@ -33,6 +33,7 @@ func Build() *cobra.Command {
 	var pipelineDir string
 	var sourceDir string
 	var signingKey string
+	var generateIndex bool
 	var useProot bool
 	var emptyWorkspace bool
 	var outDir string
@@ -56,6 +57,7 @@ func Build() *cobra.Command {
 				build.WithWorkspaceDir(workspaceDir),
 				build.WithPipelineDir(pipelineDir),
 				build.WithSigningKey(signingKey),
+				build.WithGenerateIndex(generateIndex),
 				build.WithUseProot(useProot),
 				build.WithEmptyWorkspace(emptyWorkspace),
 				build.WithOutDir(outDir),
@@ -92,6 +94,7 @@ func Build() *cobra.Command {
 	cmd.Flags().StringVar(&pipelineDir, "pipeline-dir", "/usr/share/melange/pipelines", "directory used to store defined pipelines")
 	cmd.Flags().StringVar(&sourceDir, "source-dir", "", "directory used for included sources")
 	cmd.Flags().StringVar(&signingKey, "signing-key", "", "key to use for signing")
+	cmd.Flags().BoolVar(&generateIndex, "generate-index", true, "whether to generate APKINDEX.tar.gz")
 	cmd.Flags().BoolVar(&useProot, "use-proot", false, "whether to use proot for fakeroot")
 	cmd.Flags().BoolVar(&emptyWorkspace, "empty-workspace", false, "whether the build workspace should be empty")
 	cmd.Flags().StringVar(&outDir, "out-dir", filepath.Join(cwd, "packages"), "directory where packages will be output")
